requirements.txt
shapesinc
discord.py
jishaku
tortoise-orm
flask
asyncpg
